+++ src/components/add-topics.js
@@ -1,11 +1,11 @@
 import React from 'react'
 import { InputGroup, Button, FormControl } from 'react-bootstrap'
 
-const AddTopic = () => {
+const AddTopics = () => {
   return (
     <div>
       <InputGroup className="mb-2">
-        <FormControl placeholder="Add topic..." />
+        <FormControl placeholder="Add topic... (still in construction)" />
         <InputGroup.Prepend>
           <Button type="submit" variant="secondary">
             <i className="fa fa-paper-plane-o" aria-hidden="true"></i>
@@ -16,4 +16,4 @@ const AddTopic = () => {
   )
 }
 
-export default AddTopic
+export default AddTopics
